Move `cycle_detection` module into `directed`
Since it uses the notion of successor, this is a directed algorithm.
