+++ src/directed/cycle_detection.rs
