+++ src/directed/mod.rs
@@ -6,6 +6,7 @@ use std::hash::Hash;
 pub mod astar;
 pub mod bfs;
 pub mod count_paths;
+pub mod cycle_detection;
 pub mod dfs;
 pub mod dijkstra;
 pub mod edmonds_karp;
