+++ src/lib.rs
@@ -4,9 +4,9 @@
 #![allow(clippy::module_name_repetitions)]
 #![doc = include_str!("../README.md")]
 
+use deprecate_until::deprecate_until;
 pub use num_traits;
 
-pub mod cycle_detection;
 pub mod directed;
 pub mod grid;
 pub mod kuhn_munkres;
@@ -23,10 +23,10 @@ type FxIndexSet<K> = IndexSet<K, BuildHasherDefault<FxHasher>>;
 
 /// Export all public functions and structures for an easy access.
 pub mod prelude {
-    pub use crate::cycle_detection::*;
     pub use crate::directed::astar::*;
     pub use crate::directed::bfs::*;
     pub use crate::directed::count_paths::*;
+    pub use crate::directed::cycle_detection::*;
     pub use crate::directed::dfs::*;
     pub use crate::directed::dijkstra::*;
     pub use crate::directed::edmonds_karp::*;
@@ -42,4 +42,14 @@ pub mod prelude {
     pub use crate::undirected::connected_components::*;
     pub use crate::undirected::kruskal::*;
     pub use crate::utils::*;
+}
+
+/// Deprecated: moved into the `directed` module.
+#[deprecate_until(
+    note = "use directed::cycle_detection or the prelude instead",
+    since = "4.3.1",
+    remove = "> 4.x"
+)]
+pub mod cycle_detection {
+    pub use crate::directed::cycle_detection::*;
 }
